Line too long and dead URL
Using the YARD tag syntax we can bypass the line length issue here.

However, the URL to upstream documentation is dead. Thanks a bunch
Perforce for burning old documentation.

I couldn't find a section that specifically showed how to configure
trocla, so I took something that was possibly the closest. But it's not
a perfect fit. We do have information about this in the README file
though, so we can point folks to that, too.